% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_geno}
\alias{filter_geno}
\title{Filter genotypes}
\usage{
filter_geno(candidate, geno, anno = NULL)
}
\arguments{
\item{candidate}{A data frame containing feature IDs and SNP IDs.
The first column must contain feature IDs and the second column must
contain SNP IDs.}

\item{geno}{A data frame formatted as input for MatrixEQTL. The
first column must contain SNP IDs. The rest of the columns must
contain genotypes, which can be dosage or coded as (0, 1, 2). Each
column corresponds to each sample.}

\item{anno}{A data frame containing sample IDs, subject IDs, and
treatment conditions. The columns must be named as "sample",
"subject", and "condition".}
}
\value{
A list of data frames containing the following elements:
\itemize{
\item{\code{included} - A data frame containing filtered data.}
\item{\code{excluded} - A data frame containing excluded
feature-SNP pairs.}
}
}
\description{
This function takes a \code{candidate} data frame (see
\code{\link{map_qtl}}) as input and removes SNPs for which either
of the homozygotes is missing in at least one of the control and
treated conditions.
}
\examples{
candidate <- data.frame(
    feat_id = c("f1", "f2"),
    snp_id = c("g1", "g2")
)
geno <- data.frame(
    snp_id = c("g1", "g2"),
    d1 = c(0, 1), d2 = c(1, 2), d3 = c(2, 0)
)
anno <- data.frame(
    sample = paste0("s", 1:6),
    subject = rep(c("d1", "d2", "d3"), each = 2),
    condition = rep(c(0, 1), 3)
)
filter_geno(candidate = candidate, geno = geno, anno = anno)
}
