% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_eigen}
\alias{get_eigen}
\title{Get eigenvectors and eigenvalues of the covariance matrix}
\usage{
get_eigen(anno, kinship = NULL)
}
\arguments{
\item{anno}{A data frame containing sample IDs (character strings
or integers), subject IDs (character strings or integers), and
treatment conditions (0 or 1). The columns must be named as "sample",
"subject", and "condition".}

\item{kinship}{A matrix containing pairwise genetic relatedness
between subjects. The row and column names must match the set of
unique elements of the "subject" column in "anno" in the
corresponding order (i.e., \code{unique(anno$subject)}). This is
set to \code{NULL} by default, in which case the identity matrix is
used.}
}
\value{
A list object containing:
\itemize{
\item{\code{tU} - A matrix containing the transposed eigenvectors
of the covariance matrix.}
\item{\code{lambda} - A vector containing the eigenvalues of the
covariance matrix.}
}
}
\description{
This function computes eigenvectors and eigenvalues of the
covariance matrix, which are needed for running
\code{\link{map_qtl}} and \code{\link{map_qtl_for_each}} when
modeling a random effect. \code{kinship} must be set to the default
if a subject-specific random effect will be modeled. \code{kinship}
must be specified if a polygenic (kinship) random effect will be
modeled.
This is the same as the \code{get_tu_lambda} function from the classifygxt package.
}
\examples{
anno <- data.frame(
    sample = paste0("s", 1:20),
    subject = rep(paste0("d", 1:10), each = 2),
    condition = rep(c(0, 1), 10)
)
eig <- get_eigen(anno = anno)
}
