% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geno}
\alias{get_geno}
\title{Create a genotype matrix}
\usage{
get_geno(candidate, chromosomes, plink.prefix, plink.suffix)
}
\arguments{
\item{candidate}{A data frame containing feature IDs and SNP IDs.
The first column must contain feature IDs and the second column must
contain SNP IDs.}

\item{chromosomes}{An integer vector specifying chromosomes.}

\item{plink.prefix}{A character string specifying the prefix of
PLINK files.}

\item{plink.suffix}{A character string specifying the suffix of
PLINK files.}
}
\value{
A data frame formatted as input for MatrixEQTL. The first
column contains SNP IDs. The rest of the columns contain genotypes
coded as (0, 1, 2). Each column corresponds to each sample.
}
\description{
This function takes genotype data in PLINK format as input and
returns a \code{geno} data frame, which can be passed to
\code{\link{map_qtl}} using the \code{geno} argument. See the
package vignette for details.
}
\examples{
\dontrun{
candidate <- data.frame(feat_id = "f1", snp_id = "g1")
geno <- get_geno(
    candidate = candidate,
    chromosomes = 1,
    plink.prefix = "path/to/plink",
    plink.suffix = "bed"
)
}
}
