% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mu}
\alias{get_mu}
\title{Nonlinear mean function for GxT model}
\usage{
get_mu(g, t, b0, b1, b2, b3)
}
\arguments{
\item{g}{A vector of the genotype coded as (0, 1, 2) or the
imputation-based allelic dosage.}

\item{t}{A vector of treatment indicators (0 or 1).}

\item{b0}{A scalar specifying the intercept.}

\item{b1}{A scalar specifying the genotype effect.}

\item{b2}{A scalar specifying the treatment effect.}

\item{b3}{A scalar specifying the genotype-by-treatment interaction
effect.}
}
\value{
A numeric vector of expected log expression values.
}
\description{
This function computes the expected log expression under the nonlinear
genotype-by-treatment model used by \code{\link{map_qtl}}.
}
