% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{make_sim_data}
\alias{make_sim_data}
\title{Generate data for simulation analysis}
\usage{
make_sim_data(
  num,
  anno,
  fn,
  genotype = NULL,
  lb.maf = 0.05,
  ub.maf = 0.5,
  filter.geno = FALSE,
  sd = rep(NA, 3),
  coef = rep(NA, 3),
  b0 = 0,
  sigma = 1,
  ranef = FALSE,
  sigma.u = NULL,
  kinship = NULL,
  seed = 1,
  output = "data.frame"
)
}
\arguments{
\item{num}{A named, ordered integer vector specifying the numbers of
simulations in the eight model categories.}

\item{anno}{A data frame containing sample IDs (character strings
or integers), subject IDs (character strings or integers), and
treatment conditions (0 or 1). The columns must be named as "sample",
"subject", and "condition".}

\item{fn}{A character string specifying the function. This must be
one of "nonlinear" and "linear", corresponding to nonlinear and
linear models, respectively.}

\item{genotype}{A data frame containing the subject IDs (character
strings or integers) and the genotype coded as (0, 1, 2) or the
imputation-based allelic dosage in the first and second columns,
respectively. The columns must be named as "subject" and "g".}

\item{lb.maf}{A scalar specifying the lower bound of MAF.}

\item{ub.maf}{A scalar specifying the upper bound of MAF.}

\item{filter.geno}{A logical scalar as to whether to ensure that
homozygotes for both alleles have at least one observation in both
control and treated conditions.}

\item{sd}{A numeric vector of length three specifying the standard
deviations of the genotype, treatment, and interaction effects,
respectively. For each effect, exactly one of the corresponding
elements of \code{sd} and \code{coef} must be non-\code{NA}.}

\item{coef}{A numeric vector of length three specifying fixed values
for the genotype, treatment, and interaction effects, respectively.
For each effect, exactly one of the corresponding elements of
\code{sd} and \code{coef} must be non-\code{NA}.}

\item{b0}{A scalar specifying the intercept.}

\item{sigma}{A scalar specifying the residual error standard deviation.}

\item{ranef}{A logical scalar as to whether to include random
effect.}

\item{sigma.u}{A scalar specifying the random intercept standard
deviation. If \code{ranef} is \code{TRUE}, this is set to 1
by default.}

\item{kinship}{A matrix containing pairwise genetic relatedness
between subjects. The row and column names must match the set of
unique elements of the "subject" column in "anno" in the
corresponding order (i.e., \code{unique(anno$subject)}). This is
set to \code{NULL} by default, in which case the identity matrix is
used.}

\item{seed}{An integer specifying a seed for RNG.}

\item{output}{A character string specifying the output format. This
must be one of "list" and "data.frame".}
}
\value{
A list of lists containing:
\itemize{
\item{\code{y} - A vector of phenotypes.}
\item{\code{g} - A vector of genotypes.}
\item{\code{t} - A vector of treatment indicators.}
\item{\code{sample} - A vector of sequencing sample IDs.}
\item{\code{subject} - A vector of subject IDs.}
\item{\code{index} - An integer specifying one of the eight
models.}
\item{\code{maf} - A scalar specifying the minor allele frequency
used for generating the genotype data.}
\item{\code{beta} - A named numeric vector specifying the true
coefficient values used for generating the phenotype data. The
"b0" element represents the intercept. The "b1", "b2", and "b3"
elements respectively represent the genotype, treatment, and
interaction effect sizes. }
} or a list containing:
\itemize{
\item{\code{candidate} - A data frame containing feature IDs and
SNP IDs. See \code{\link{map_qtl}}. This also includes the integers
specifying models, the minor allele frequency, and the true
coefficient values.}
\item{\code{geno} - A data frame containing genotypes. See \code{\link{map_qtl}}.}
\item{\code{pheno} - A data frame containing phenotypes. See \code{\link{map_qtl}}.}
}
}
\description{
This function generates simulated genotype and phenotype data for
all eight GxT model categories.
}
\examples{
anno <- data.frame(
    sample = paste0("s", 1:20),
    subject = rep(paste0("d", 1:10), each = 2),
    condition = rep(c(0, 1), 10)
)
dat <- make_sim_data(
    num = c(0, 0, 0, 0, 0, 0, 0, 1),
    anno = anno,
    fn = "nonlinear",
    coef = c(0.5, 0.3, 0.2)
)
}
