% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_pheno}
\alias{preprocess_pheno}
\title{Pre-process molecular count data}
\usage{
preprocess_pheno(count, anno, covar = NULL, num.pc = 10, list = FALSE)
}
\arguments{
\item{count}{A matrix containing molecular counts (non-negative
integers). The row and column names correspond to feature IDs and
sample IDs, respectively.}

\item{anno}{A data frame containing sample IDs, subject IDs, and
treatment conditions. The columns must be named as "sample",
"subject", and "condition".}

\item{covar}{A data frame containing fixed effect covariates. The
first column must contain sample IDs. The rest of the columns must
contain covariate values. Each row corresponds to each sample. Each
column corresponds to each covariate.}

\item{num.pc}{An integer specifying the number of molecular
phenotype principal components to control for.}

\item{list}{A logical scalar. If \code{TRUE}, the PCA result,
i.e., the output from the \code{prcomp} function in the stats
package, will be returned as an element of a list.}
}
\value{
A data frame containing pre-processed phenotype data or a
list containing the following elements:
\itemize{
\item{\code{pheno} - A data frame containing pre-processed
phenotype data.}
\item{\code{pca} - A list object from the \code{prcomp} function in
the stats package.}
}
}
\description{
This function takes raw molecular count data as input and returns a
\code{pheno} data frame, which can be passed to
\code{\link{map_qtl}} using the \code{pheno} argument.
}
\examples{
anno <- data.frame(
    sample = paste0("s", 1:20),
    subject = rep(paste0("d", 1:10), each = 2),
    condition = rep(c(0, 1), 10)
)
count <- matrix(
    rpois(100, lambda = 100), nrow = 5,
    dimnames = list(paste0("f", 1:5), paste0("s", 1:20))
)
pheno <- preprocess_pheno(count = count, anno = anno, num.pc = 3)
}
